Do not parse --network when --pod is set
Always parsing the --network flag does not work since it overwrites the
pod netns when --pod is given to `podman create/run` command. Parse
the --network value only when it is explicity given on the command line
or when --pod is not used. With `podman pod create` do not parse it when
no infra container is created.

This still does not work when a pod without shared netns is used. In
this case we will use the default netns instead of the one from the
config file. Making this work requires reworking the parsing logic and
moving it into specgen.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>